API All For: employee, department, position, role

+++ src/main/java/com/comaymanagement/cmd/repository/IDepartmentRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.comaymanagement.cmd.entity.Department;
 
 
-public interface IDepartmentRepository extends JpaRepository<Department, Long>{
+public interface IDepartmentRepository extends JpaRepository<Department, Integer>{
 	
 	@Query(value = "Select * FROM cmd.departments dp inner join cmd.departments_employees de "
 			+ "on dp.id = de.department_id "
